# backend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Mirzagram API
  version: 1.0.0
  description: APIs for Mirzagram.
...
security:
  - BearerAuth:
      - read
      - write
paths:
...
  /post:
    get:
      summary: Get all posts for user
      tags: [Post]
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#components/schemas/ApiResponse"
                  - type: object
                    properties:
                      result:
                        $ref: "#components/schemas/PostSummaryDTO"
    post:
      summary: Create new post
      tags: [Post]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequestDTO"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#components/schemas/ApiResponse"
                  - type: object
                    properties:
                      result:
                        type: string
  /post/{id}:
    get:
      summary: Get post details
      tags: [Post]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The id of the post to retrieve
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#components/schemas/ApiResponse"
                  - type: object
                    properties:
                      result:
                        $ref: "#components/schemas/PostDTO"
    post:
      summary: Edit existing post
      tags: [Post]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePostRequestDTO"
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: The id of the post to edit
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#components/schemas/ApiResponse"
                  - type: object
                    properties:
                      result:
                        type: string
  /post/like:
    get:
      summary: Get likes for post
      tags: [Post]
      parameters:
        - in: query
          name: postId
          schema:
            type: integer
          required: true
          description: Post id
        - in: query
          name: page
          schema:
            type: integer
          required: true
          description: Page number
        - in: query
          name: pageSize
          schema:
            type: integer
          required: true
          description: Page size
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#components/schemas/ApiResponse"
                  - type: object
                    properties:
                      result:
                        $ref: "#components/schemas/PostLikesResponseDTO"
    post:
      summary: Like a post
      tags: [Post]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostLikeRequestDTO"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#components/schemas/ApiResponse"
                  - type: object
    delete:
      summary: Unlike a post
      tags: [Post]
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostLikeRequestDTO"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#components/schemas/ApiResponse"
                  - type: object
  # Comment
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    CreatePostRequestDTO:
      type: object
      required:
        - fileNames
        - caption
        - mentions
      properties:
        fileNames:
          type: array
          items:
            type: string
        caption:
          type: string
        mentions:
          type: array
          items:
            type: string
...
    PostLikeRequestDTO:
      type: object
      required:
        - postId
      properties:
        postId:
          type: string
...